sava checkpoint update

+++ src/configs/stage1_config.py
@@ -156,7 +156,7 @@ parser.add_argument(
 parser.add_argument(
     "--mixed_precision",
     type=str,
-    default="fp16",
+    default="bf16",
     choices=["no", "fp16", "bf16"],
     help=(
         "Whether to use mixed precision. Choose between fp16 and bf16 (bfloat16). Bf16 requires PyTorch >="
